requirements.txt
easyocr==1.6.2
torch==2.0.1
torchvision==0.15.2
Pillow==9.5.0
numpy==1.26.4
pdf2image==1.17.0
gradio==4.44.1
pydantic==2.6.4
transformers==4.38.2
scikit-image==0.20.0
opencv-python-headless==4.7.0.72